Fix energy axis for DM spectra. There might be some edge cases where this is still incorrect.
dmreader: Some single spectra are returned as 2D with a singular dimension for older spectrometer data. The logic to remove the extra dimension is difficult due to conflicts with newer data sets from modern spectrometers.

dmreader: Remove the confusing pixelOrigin key in the dictionary. Use the coords instead.

Add note regarding the pixelOrigin in getDataset to docstring.

Update some formatting and docstrings.